pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "xsarena"
version = "0.2.0"
description = "AI-powered writing and coding studio"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "XSArena Team"}
]
dependencies = [
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "requests>=2.31.0",
    "jsonschema>=4.18.0"
]

[project.optional-dependencies]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0"
]
watch = [
    "watchdog>=3.0.0"
]
metrics = [
    "prometheus-client>=0.19.0"
]


[project.scripts]
xsarena = "xsarena.cli.main:run"
# xsarena-tui = "xsarena_tui:main"   # moved to contrib/; no longer maintained as core feature
xsarena-bridge = "xsarena.bridge_v2.api_server:run_server"
# compatibility for one release cycle:
lmastudio = "xsarena.cli.main:run"
lmastudio-bridge = "xsarena.bridge_v2.api_server:run_server"

[project.urls]
Homepage = "https://github.com/your-username/xsarena"
Repository = "https://github.com/your-username/xsarena"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
